Create Tradingstrategies
Making the socio-economic rights judicially enforceable will create an obligation on the state to fulfil them and utilize the resources well to bring an equitable atmosphere for its citizens and reduce the imbalance in the society. It is necessary because in many third world countries of the world, economic growth has happened but the living standard of the people has not improved. Bangladesh has also ratified the ICESCR which imposes an obligation to make these socio-economic rights justiciable in its territorial jurisdiction as much as it can according to its availability of resources. The government has to advance with proper planning to execute these rights as the territory of Bangladesh is not too big for its gigantic population.